admin-ui/requirements.txt
streamlit
asyncio
websockets
pandas
psycopg2-binary
wordcloud
matplotlib
plotly
openapi-spec-validator  
pymupdf
openpyxl
openapi
python-pptx